+++ cip25/rust/Cargo.toml
@@ -8,11 +8,14 @@ keywords = ["cardano"]
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
+cml-chain = { "path" = "../../chain/rust" }
 cml-core = { "path" = "../../core/rust" }
 cml-crypto = { "path" = "../../crypto/rust" }
-cml-chain = { "path" = "../../chain/rust" }
 cbor_event = "2.2.0"
+hex = "0.4.0"
+schemars = "0.8.8"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.57"
-schemars = "0.8.8"
-hex = "0.4.0"
+thiserror = "1.0.37"
+# for enums
+wasm-bindgen = { version = "=0.2.83", features = ["serde-serialize"] }
