+++ cip25/wasm/Cargo.toml
@@ -8,10 +8,13 @@ keywords = ["cardano"]
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-core = { path = "../rust", package = "cml-cip25" }
-cml-core-wasm = { path = "../../core/wasm" }
 cbor_event = "2.2.0"
-wasm-bindgen = { version = "=0.2.83", features = ["serde-serialize"] }
+cml-chain = { path = "../../chain/rust" }
+cml-chain-wasm = { path = "../../chain/wasm" }
+cml-core = { path = "../../core/rust" }
+cml-core-wasm = { path = "../../core/wasm" }
+core = { path = "../rust", package = "cml-cip25" }
 linked-hash-map = "0.5.3"
 serde_json = "1.0.57"
 serde-wasm-bindgen = "0.4.5"
+wasm-bindgen = { version = "=0.2.83", features = ["serde-serialize"] }
